requirements.txt
PyYAML
argparse
matplotlib
numpy
scipy
torch==1.10.1
torchvision==0.11.2